chore: extract postgres-changes suite into src/suites/postgres-changes.ts
Part of REAL-1037.
